test=huawei_ascend_npu test=verisilicon_timvx test=nvidia_tensorrt
