+++ lite/core/optimizer/mir/nnadapter_restrict_quantized_op_with_same_input_output_scale_pass.h
@@ -14,13 +14,8 @@
 
 #pragma once
 
-#include <limits>
-#include <map>
 #include <memory>
-#include <string>
-#include <vector>
 #include "lite/core/optimizer/mir/pass.h"
-#include "lite/core/types.h"
 
 namespace paddle {
 namespace lite {
@@ -73,7 +68,8 @@ namespace mir {
 //                           |
 //                           v
 
-class RestrictQuantizedOpWithSameInputOutputScalePass : public mir::StmtPass {
+class NNAdapterRestrictQuantizedOpWithSameInputOutputScalePass
+    : public mir::StmtPass {
  public:
   void Apply(const std::unique_ptr<SSAGraph>& graph) override;
 };
